Only kill of pidfile is readable integer

+++ lib/redis_test.rb
@@ -74,8 +74,10 @@ module RedisTest
     end
 
     def stop
-      pid = File.read(pidfile).to_i
-      Process.kill("QUIT", pid)
+      if File.file?(pidfile) && File.readable?(pidfile)
+        pid = File.read(pidfile).to_i
+        Process.kill("QUIT", pid) if pid > 0
+      end
       FileUtils.rm_f("#{cache_path}#{db_filename}")
     end
